Re-import audio clips (grr)
The 5.1.3 editor refuses to treat WAV files in StreamingAssets
as imports.  This means you can't drop them into the AudioClip
field of a Unity AudioSource object.  The problem can be worked
around by importing the WAV files in a different directory,
configuring them, and then dragging them into the StreamingAssets
folder.

The AudioClip aspect will survive editor restarts, but will be
lost if the Library folder is regenerated.

This change does nothing but update the GUIDs in my project, which
change with each import.